Refactor runner.rs
