+++ wireguard-rs/src/run/mod.rs
@@ -4,5 +4,6 @@ mod config;
 mod profiler;
 mod runner;
 mod util;
+mod workers;
 
 pub use runner::create_config_and_run;
