+++ wireguard-rs/src/run/workers/uapi.rs
@@ -1,0 +1,137 @@
+use std::io::{self, BufReader, BufWriter, Write};
+use std::process::exit;
+use std::thread::{self, JoinHandle, ScopedJoinHandle};
+
+use wg_platform as plt;
+use wg_traits::{
+    tun::Tun,
+    uapi::{BindUAPI, PlatformUAPI},
+    udp::PlatformUDP,
+};
+
+use crate::configuration::{ConfigError, Configuration, WireGuardConfig, uapi};
+use crate::run::{error::ExitCode, profiler::profiler_stop};
+use crate::wireguard::WireGuard;
+
+pub enum ConfigMessage {
+    UapiConfigOperation(
+        uapi::ConfigOperation,
+        crossbeam_channel::Sender<Result<String, ConfigError>>,
+    ),
+    TunUp(usize),
+    TunDown,
+}
+
+pub fn spawn_config_worker<'scope, 'env, T: Tun, B: PlatformUDP>(
+    thread_scope: &'scope thread::Scope<'scope, 'env>,
+    wireguard_device: &'env WireGuard<T, B>,
+    config_receiver: crossbeam_channel::Receiver<ConfigMessage>,
+) -> ScopedJoinHandle<'scope, ()> {
+    thread_scope.spawn(|| {
+        let mut wireguard_config = WireGuardConfig::new(wireguard_device.clone());
+        let config_receiver = config_receiver;
+
+        while let Ok(message) = config_receiver.recv() {
+            match message {
+                ConfigMessage::UapiConfigOperation(config_operation, sender) => {
+                    sender
+                        .send(uapi::handle_config_operation(
+                            config_operation,
+                            &mut wireguard_config,
+                        ))
+                        .expect("channel is open until result is received");
+                }
+                ConfigMessage::TunUp(mtu) => {
+                    log::info!("Tun up (mtu = {})", mtu);
+                    if let Err(err) = wireguard_config.up(mtu) {
+                        log::error!("Error during TUN setup: {err}");
+                    }
+                }
+                ConfigMessage::TunDown => {
+                    log::info!("Tun down");
+                    wireguard_config.down();
+                }
+            }
+        }
+    })
+}
+
+pub fn spawn_uapi_server(
+    uapi: <plt::UAPI as PlatformUAPI>::Bind,
+    config_sender: crossbeam_channel::Sender<ConfigMessage>,
+) -> JoinHandle<()> {
+    thread::spawn(move || {
+        loop {
+            // accept and handle UAPI config connections
+            match uapi.connect() {
+                Ok(stream) => {
+                    spawn_uapi_config_message_handler(config_sender.clone(), stream);
+                }
+                Err(err) => {
+                    log::error!("UAPI connection error: {}", err);
+                    profiler_stop();
+                    exit(ExitCode::UAPIConnectionError as i32);
+                }
+            }
+        }
+    })
+}
+
+fn spawn_uapi_config_message_handler(
+    config_sender: crossbeam_channel::Sender<ConfigMessage>,
+    mut stream: <<plt::UAPI as PlatformUAPI>::Bind as BindUAPI>::Stream,
+) -> JoinHandle<()> {
+    thread::spawn(move || {
+        let mut reader = BufReader::new(&mut stream);
+        let mut string_buffer = String::new();
+
+        'read_from_stream: loop {
+            let result = uapi::parse_config_operation(&mut reader, &mut string_buffer).and_then(
+                |config_operation| match config_operation {
+                    Some(config_operation) => {
+                        let (sender, receiver) = crossbeam_channel::unbounded();
+
+                        config_sender
+                            .send(ConfigMessage::UapiConfigOperation(config_operation, sender))
+                            .expect("channel is open while this loop is running");
+
+                        receiver
+                            .recv()
+                            .expect("channel is open until result is received")
+                            .map(Some)
+                    }
+                    None => Ok(None), // channel closed
+                },
+            );
+
+            if let Ok(None) = result {
+                // channel closed
+                break 'read_from_stream;
+            }
+
+            let mut errno = 0;
+
+            let response = match result {
+                Ok(response) => response.expect("None case was already handled"),
+                Err(err) => {
+                    log::error!("Error during config operation: {err}");
+
+                    errno = err.errno();
+
+                    String::new()
+                }
+            };
+
+            if let Err(err) = || -> io::Result<_> {
+                let mut writer = BufWriter::new(reader.get_mut());
+
+                writer.write_all(response.as_bytes())?;
+                writer.write_all(format!("errno={errno}\n\n").as_bytes())?;
+
+                Ok(())
+            }() {
+                log::error!("Error while writing to Unix socket: {err}");
+            }
+        }
+    })
+}
